+++ dpdk/src/port.rs
@@ -32,16 +32,16 @@ impl RxOffload {
     pub fn to_flags(&self) -> u64 {
         let mut flags = 0u64;
         if self.vlan_strip {
-            flags |= dpdk_sys::RTE_ETH_RX_OFFLOAD_VLAN_STRIP;
+            flags |= dpdk_sys::RTE_ETH_RX_OFFLOAD_VLAN_STRIP as u64;
         }
         if self.ipv4_cksum {
-            flags |= dpdk_sys::RTE_ETH_RX_OFFLOAD_IPV4_CKSUM;
+            flags |= dpdk_sys::RTE_ETH_RX_OFFLOAD_IPV4_CKSUM as u64;
         }
         if self.udp_cksum {
-            flags |= dpdk_sys::RTE_ETH_RX_OFFLOAD_UDP_CKSUM;
+            flags |= dpdk_sys::RTE_ETH_RX_OFFLOAD_UDP_CKSUM as u64;
         }
         if self.tcp_cksum {
-            flags |= dpdk_sys::RTE_ETH_RX_OFFLOAD_TCP_CKSUM;
+            flags |= dpdk_sys::RTE_ETH_RX_OFFLOAD_TCP_CKSUM as u64;
         }
         flags
     }
@@ -65,16 +65,16 @@ impl TxOffload {
     pub fn to_flags(&self) -> u64 {
         let mut flags = 0u64;
         if self.vlan_insert {
-            flags |= dpdk_sys::RTE_ETH_TX_OFFLOAD_VLAN_INSERT;
+            flags |= dpdk_sys::RTE_ETH_TX_OFFLOAD_VLAN_INSERT as u64;
         }
         if self.ipv4_cksum {
-            flags |= dpdk_sys::RTE_ETH_TX_OFFLOAD_IPV4_CKSUM;
+            flags |= dpdk_sys::RTE_ETH_TX_OFFLOAD_IPV4_CKSUM as u64;
         }
         if self.udp_cksum {
-            flags |= dpdk_sys::RTE_ETH_TX_OFFLOAD_UDP_CKSUM;
+            flags |= dpdk_sys::RTE_ETH_TX_OFFLOAD_UDP_CKSUM as u64;
         }
         if self.tcp_cksum {
-            flags |= dpdk_sys::RTE_ETH_TX_OFFLOAD_TCP_CKSUM;
+            flags |= dpdk_sys::RTE_ETH_TX_OFFLOAD_TCP_CKSUM as u64;
         }
         flags
     }
@@ -292,32 +292,32 @@ pub struct DeviceCapabilities {
 impl DeviceCapabilities {
     /// Check if RX IPv4 checksum offload is supported
     pub fn supports_rx_ipv4_cksum(&self) -> bool {
-        (self.rx_offload_capa & dpdk_sys::RTE_ETH_RX_OFFLOAD_IPV4_CKSUM) != 0
+        (self.rx_offload_capa & dpdk_sys::RTE_ETH_RX_OFFLOAD_IPV4_CKSUM as u64) != 0
     }
 
     /// Check if RX UDP checksum offload is supported
     pub fn supports_rx_udp_cksum(&self) -> bool {
-        (self.rx_offload_capa & dpdk_sys::RTE_ETH_RX_OFFLOAD_UDP_CKSUM) != 0
+        (self.rx_offload_capa & dpdk_sys::RTE_ETH_RX_OFFLOAD_UDP_CKSUM as u64) != 0
     }
 
     /// Check if TX IPv4 checksum offload is supported
     pub fn supports_tx_ipv4_cksum(&self) -> bool {
-        (self.tx_offload_capa & dpdk_sys::RTE_ETH_TX_OFFLOAD_IPV4_CKSUM) != 0
+        (self.tx_offload_capa & dpdk_sys::RTE_ETH_TX_OFFLOAD_IPV4_CKSUM as u64) != 0
     }
 
     /// Check if TX UDP checksum offload is supported
     pub fn supports_tx_udp_cksum(&self) -> bool {
-        (self.tx_offload_capa & dpdk_sys::RTE_ETH_TX_OFFLOAD_UDP_CKSUM) != 0
+        (self.tx_offload_capa & dpdk_sys::RTE_ETH_TX_OFFLOAD_UDP_CKSUM as u64) != 0
     }
 
     /// Check if VLAN stripping is supported
     pub fn supports_vlan_strip(&self) -> bool {
-        (self.rx_offload_capa & dpdk_sys::RTE_ETH_RX_OFFLOAD_VLAN_STRIP) != 0
+        (self.rx_offload_capa & dpdk_sys::RTE_ETH_RX_OFFLOAD_VLAN_STRIP as u64) != 0
     }
 
     /// Check if VLAN insertion is supported
     pub fn supports_vlan_insert(&self) -> bool {
-        (self.tx_offload_capa & dpdk_sys::RTE_ETH_TX_OFFLOAD_VLAN_INSERT) != 0
+        (self.tx_offload_capa & dpdk_sys::RTE_ETH_TX_OFFLOAD_VLAN_INSERT as u64) != 0
     }
 }
